Move CaretPositionChanged.
